% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi3.R
\name{bmi3}
\alias{bmi3}
\title{Create BMI Groups}
\usage{
bmi3(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Factor variable.
}
\description{
Converts continuous BMI variable to BMI groups.
}
\examples{
bmi.vals <- rnorm(n = 50, mean = 25, sd = 3)
bmi3(bmi.vals)

}
