% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATH4753butl3043.R
\docType{package}
\name{MATH4753butl3043}
\alias{MATH4753butl3043}
\title{BMI Package}
\description{
Contains one function for creating BMI groups.
}
\author{
Sydney Butler \email{sydney.n.butler-1@ou.edu}
}
